Implements authentication with reajc-js :lock:
