+++ todoapp/public/keycloak.json
@@ -1,0 +1,9 @@
+{
+  "realm": "todoapp",
+  "auth-server-url": "http://localhost:8080/auth/",
+  "ssl-required": "external",
+  "resource": "todo-app",
+  "public-client": true,
+  "verify-token-audience": true,
+  "confidential-port": 0
+}
