+++ todoapp/src/components/Welcome.js
@@ -1,0 +1,12 @@
+
+import React from 'react';
+
+const Welcome = () => {
+    return (
+        <div className="welcome" >
+            <p>Welcome</p>
+        </div>
+    );
+};
+
+export default Welcome;
